fix: Change the data types of decimal properties #3

+++ src/WebApp/Context/Drink.cs
@@ -10,7 +10,7 @@ namespace WebApp.Context
         public string Name { get; set; }
         public int Quantity { get; set; }
         public string Description { get; set; }
-        public double OverheadCost { get; set; }
+        public decimal OverheadCost { get; set; }
 
         public DateTime Created { get; set; }
         public DateTime? Modified { get; set; }
